Correctly refactor tests
