+++ test/plugins/test_albumtypes.py
