+++ test/plugins/test_art.py
