+++ test/plugins/test_bareasc.py
