+++ test/plugins/test_discogs.py
