+++ test/plugins/test_filefilter.py
