+++ test/plugins/test_hook.py
