+++ test/plugins/test_ihate.py
