+++ test/plugins/test_importadded.py
