+++ test/plugins/test_info.py
