+++ test/plugins/test_keyfinder.py
