+++ test/plugins/test_lyrics.py
