+++ test/plugins/test_mpdstats.py
