+++ test/plugins/test_play.py
