+++ test/plugins/test_player.py
