+++ test/plugins/test_playlist.py
