+++ test/plugins/test_smartplaylist.py
