+++ test/plugins/test_spotify.py
