+++ test/plugins/test_subsonicupdate.py
