+++ test/plugins/test_thumbnails.py
